% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runtimeconfig_objects.R
\name{Cardinality}
\alias{Cardinality}
\title{Cardinality Object}
\usage{
Cardinality(path = NULL, number = NULL)
}
\arguments{
\item{path}{The root of the variable subtree to monitor}

\item{number}{The number variables under the `path` that must exist to meet this}
}
\value{
Cardinality object
}
\description{
Cardinality Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A Cardinality condition for the Waiter resource. A cardinality condition ismet when the number of variables under a specified path prefix reaches apredefined number. For example, if you set a Cardinality condition wherethe `path` is set to `/foo` and the number of paths is set to 2, thefollowing variables would meet the condition in a RuntimeConfig resource:+ `/foo/variable1 = 'value1'`+ `/foo/variable2 = 'value2'`+ `/bar/variable3 = 'value3'`It would not would not satisify the same condition with the `number` set to3, however, because there is only 2 paths that start with `/foo`.Cardinality conditions are recursive; all subtrees under the specificpath prefix are counted.
}
