% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runtimeconfig_objects.R
\name{EndCondition}
\alias{EndCondition}
\title{EndCondition Object}
\usage{
EndCondition(cardinality = NULL)
}
\arguments{
\item{cardinality}{The cardinality of the `EndCondition`}
}
\value{
EndCondition object
}
\description{
EndCondition Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The condition that a Waiter resource is waiting for.
}

