% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runtimeconfig_objects.R
\name{ListConfigsResponse}
\alias{ListConfigsResponse}
\title{ListConfigsResponse Object}
\usage{
ListConfigsResponse(nextPageToken = NULL, configs = NULL)
}
\arguments{
\item{nextPageToken}{This token allows you to get the next page of results for list requests}

\item{configs}{A list of the configurations in the project}
}
\value{
ListConfigsResponse object
}
\description{
ListConfigsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
`ListConfigs()` returns the following response. The order of returnedobjects is arbitrary; that is, it is not ordered in any particular way.
}

