% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runtimeconfig_objects.R
\name{ListVariablesResponse}
\alias{ListVariablesResponse}
\title{ListVariablesResponse Object}
\usage{
ListVariablesResponse(variables = NULL, nextPageToken = NULL)
}
\arguments{
\item{variables}{A list of variables and their values}

\item{nextPageToken}{This token allows you to get the next page of results for list requests}
}
\value{
ListVariablesResponse object
}
\description{
ListVariablesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response for the `ListVariables()` method.
}

