% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runtimeconfig_objects.R
\name{ListWaitersResponse}
\alias{ListWaitersResponse}
\title{ListWaitersResponse Object}
\usage{
ListWaitersResponse(nextPageToken = NULL, waiters = NULL)
}
\arguments{
\item{nextPageToken}{This token allows you to get the next page of results for list requests}

\item{waiters}{Found waiters in the project}
}
\value{
ListWaitersResponse object
}
\description{
ListWaitersResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response for the `ListWaiters()` method.Order of returned waiter objects is arbitrary.
}
