% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runtimeconfig_functions.R
\name{projects.configs.create}
\alias{projects.configs.create}
\title{Creates a new RuntimeConfig resource. The configuration name must beunique within project.}
\usage{
projects.configs.create(RuntimeConfig, parent, requestId = NULL)
}
\arguments{
\item{RuntimeConfig}{The \link{RuntimeConfig} object to pass to this method}

\item{parent}{The [project ID](https://support}

\item{requestId}{An optional but recommended unique `request_id`}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/cloudruntimeconfig
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/cloudruntimeconfig)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/deployment-manager/runtime-configurator/}{Google Documentation}

Other RuntimeConfig functions: \code{\link{RuntimeConfig}},
  \code{\link{projects.configs.update}}
}
