% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runtimeconfig_functions.R
\name{projects.configs.get}
\alias{projects.configs.get}
\title{Gets information about a RuntimeConfig resource.}
\usage{
projects.configs.get(name)
}
\arguments{
\item{name}{The name of the RuntimeConfig resource to retrieve, in the format:}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/cloudruntimeconfig
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/cloudruntimeconfig)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/deployment-manager/runtime-configurator/}{Google Documentation}
}
