% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runtimeconfig_functions.R
\name{projects.configs.list}
\alias{projects.configs.list}
\title{Lists all the RuntimeConfig resources within project.}
\usage{
projects.configs.list(parent, pageSize = NULL, pageToken = NULL)
}
\arguments{
\item{parent}{The [project ID](https://support}

\item{pageSize}{Specifies the number of results to return per page}

\item{pageToken}{Specifies a page token to use}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/cloudruntimeconfig
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/cloudruntimeconfig)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/deployment-manager/runtime-configurator/}{Google Documentation}
}

