% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runtimeconfig_functions.R
\name{projects.configs.update}
\alias{projects.configs.update}
\title{Updates a RuntimeConfig resource. The configuration must exist beforehand.}
\usage{
projects.configs.update(RuntimeConfig, name)
}
\arguments{
\item{RuntimeConfig}{The \link{RuntimeConfig} object to pass to this method}

\item{name}{The name of the RuntimeConfig resource to update, in the format:}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/cloudruntimeconfig
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/cloudruntimeconfig)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/deployment-manager/runtime-configurator/}{Google Documentation}

Other RuntimeConfig functions: \code{\link{RuntimeConfig}},
  \code{\link{projects.configs.create}}
}

