% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runtimeconfig_objects.R
\name{RuntimeConfig}
\alias{RuntimeConfig}
\title{RuntimeConfig Object}
\usage{
RuntimeConfig(description = NULL, name = NULL)
}
\arguments{
\item{description}{An optional description of the RuntimeConfig object}

\item{name}{The resource name of a runtime config}
}
\value{
RuntimeConfig object
}
\description{
RuntimeConfig Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A RuntimeConfig resource is the primary resource in the Cloud RuntimeConfigservice. A RuntimeConfig resource consists of metadata and a hierarchy ofvariables.
}
\seealso{
Other RuntimeConfig functions: \code{\link{projects.configs.create}},
  \code{\link{projects.configs.update}}
}

