% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runtimeconfig_objects.R
\name{Variable}
\alias{Variable}
\title{Variable Object}
\usage{
Variable(value = NULL, text = NULL, updateTime = NULL, state = NULL,
  name = NULL)
}
\arguments{
\item{value}{The binary value of the variable}

\item{text}{The textual value of the variable}

\item{updateTime}{[Output Only] The time of the last variable update}

\item{state}{[Ouput only] The current state of the variable}

\item{name}{The name of the variable resource, in the format:}
}
\value{
Variable object
}
\description{
Variable Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Describes a single variable within a RuntimeConfig resource.The name denotes the hierarchical variable name. For example,`ports/serving_port` is a valid variable name. The variable value is anopaque string and only leaf variables can have values (that is, variablesthat do not have any child variables).
}

