% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runtimeconfig_objects.R
\name{WatchVariableRequest}
\alias{WatchVariableRequest}
\title{WatchVariableRequest Object}
\usage{
WatchVariableRequest(newerThan = NULL)
}
\arguments{
\item{newerThan}{If specified, checks the current timestamp of the variable and if the}
}
\value{
WatchVariableRequest object
}
\description{
WatchVariableRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request for the `WatchVariable()` method.
}

