% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runtimeconfig_objects.R
\name{ListVariablesResponse}
\alias{ListVariablesResponse}
\title{Google Cloud RuntimeConfig API Objects 
Provides capabilities for dynamic configuration and coordination for applications running on Google Cloud Platform.}
\usage{
ListVariablesResponse(variables = NULL, nextPageToken = NULL)
}
\arguments{
\item{variables}{A list of variables and their values}

\item{nextPageToken}{This token allows you to get the next page of results for list requests}
}
\value{
ListVariablesResponse object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 20:11:18
filename: /Users/mark/dev/R/autoGoogleAPI/googleruntimeconfigv1beta1.auto/R/runtimeconfig_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
ListVariablesResponse Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response for the `ListVariables()` method.
}
